+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## dev-master
+
+* `->tmpDir()->getPath()`: clear argument (the target path must not exist)
+
 ## 0.1.0 (18.03.2023)
 
 * `BaseAxyTestCase`